feat: 🎨 createAsset refactoring
