+++ src/components/TransactionForms/CustomForms/utils/index.ts
@@ -15,6 +15,7 @@ export const parseSplitRoyalties = (data: any): void => {
     data.royalties?.splitRoyalties === undefined ||
     !data.royalties?.splitRoyalties?.length
   ) {
+    data.royalties = {};
     return;
   }
